app/requirements/production.txt
django
django-crispy-forms
gunicorn
django-filter